+++ src/main/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/AnySchema.scala
@@ -6,21 +6,26 @@ import scala.reflect._
 import org.json4s.JsonDSL._
 import org.json4s._
 
-final case class ZeroSchema(
+final case class AnySchema(
     override val properties: SchemaProperties[Nothing] = SchemaProperties.empty
 ) extends JsonSchema[Nothing] {
-  override val schemaType = "zero"
+  override val schemaType = "any"
 
-  override def toJson: JObject = ("not" -> Nil)
+  override def toJson: JObject = Nil
 
+  // XXX Actually all types are valid, but we still have to specify this set
   override val validTypes: Set[ClassTag[_ <: JValue]] = Set.empty
+
+  override def isValidType[S <: JValue](value: S)(implicit
+      tag: ClassTag[S]
+  ): Boolean = true
 
   def mergeSameType()(implicit
       er: EquivalenceRelation
   ): PartialFunction[JsonSchema[_], JsonSchema[_]] = { case _ @other =>
-    other
+    this
   }
 
-  override def copy(properties: SchemaProperties[Nothing]): ZeroSchema =
-    ZeroSchema(properties)
+  override def copy(properties: SchemaProperties[Nothing]): AnySchema =
+    AnySchema(properties)
 }